updated Mossing 12 Month report
I adjusted the report for singular participants. Tests if Participant responded to certain fields that were not populated when hard copy checked.

One field does not want to be pulled, "inf2mo_healthcare_services" and "inf2wk_healthcare_services". Most participants did not respond to this field, as such I think it's important to check why this is not pulling.