this is my BEACH Recruitment Report
There are some issues that will be layed out in the issues pannel. I'm having an issue pulling certain events from REDcap.

This code will tell you:
the particiapnts that still need to consent.
summary of encounter types.
quarterly breakdown of encounters.

There is more fuctionallity to add, but the pulling issues must be addressed first.